Extract cacheable api base into a separate abstract class
